leam-base.joda-time.mid-1186.idx-172014.5.mutant

+++ joda-time/src/main/java/org/joda/time/chrono/GJLocaleSymbols.java
@@ -172,7 +172,7 @@ class GJLocaleSymbols {
     public int eraTextToValue(String text) {
         Integer era = iParseEras.get(text);
         if (era != null) {
-            return era.intValue();
+            return era;
         }
         throw new IllegalFieldValueException(DateTimeFieldType.era(), text);
     }
